fix: delay in showing keyboard when click on keyboard icon
unfortunately the `keyboard.show()` has a delay in showing
the keyboard so I used the focus function of HTMLElement to
show keyboard. but there is notice that focus method doesn't
work when the element is already focused and causes the
keyboard to stay hidden. so I used the blur method to remove
focus from text area.